third_party/WebKit/LayoutTests/inspector/tracing/worker-js-frames.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>

// Save references to the worker objects to make sure they are not GC'ed.
var worker1;
var worker2;

// The function ensures call stack for the current frame is captured.
function makeDOMEvent()
{
    var e = document.createElement("div");
    e.innerHTML = "<p>test</p>";
    return e.lastChild.textContents;
}

function startWorkerAndRunTest()
{
    worker1 = new Worker("resources/worker.js");
    worker1.postMessage("");

    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }

    worker1.onmessage = function(event)
    {
        worker1.onmessage = null;
        runTest();
    }
}

function startSecondWorker(onActionComplete)
{
    // Make sure there is at least one trace event recorded for the first worker.
    worker1.onmessage = gotResponseFromFirstWorker;
    worker1.postMessage("");
    makeDOMEvent();

    function gotResponseFromFirstWorker()
    {
        worker1.onmessage = null;

        worker2 = new Worker("resources/worker.js");
        worker2.postMessage("");
        worker2.onmessage = function(event)
        {
            worker2.onmessage = null;
            makeDOMEvent();
            onActionComplete();
        }
    }
}

function test()
{
    InspectorTest.invokeWithTracing("startSecondWorker", processTracingEvents,
        "disabled-by-default-devtools.timeline.stack", true);

    function processTracingEvents()
    {
        var mainThread = { name: "Main Thread", events: InspectorTest.timelineModel()._mainThreadEvents };
        processThread(new Set(["worker2.onmessage"]), mainThread);
        InspectorTest.timelineModel()._virtualThreads.forEach(function(thread)
        {
            if (!thread.isWorker())
                return;
            processThread(new Set(["onmessage"]), thread);
        });
        InspectorTest.completeTest();
    }

    function processThread(expectedFunctions, thread)
    {
        InspectorTest.addResult(thread.name);
        var missingFunctions = thread.events.reduce(processEvent, expectedFunctions);
        if (missingFunctions.size) {
            InspectorTest.addResult("FAIL: missing functions:");
            missingFunctions.forEach(InspectorTest.addResult);
        }
    }

    function processEvent(expectedFunctions, event)
    {
        if (event.name === WebInspector.TimelineModel.RecordType.JSFrame)
            expectedFunctions.delete(event.args.data.functionName);
        return expectedFunctions;
    }
}

</script>
</head>

<body onload="startWorkerAndRunTest()">
<p>
Tests js cpu profile in timeline.
</p>
</body>
</html>
